[Copilot] Extract ConstraintMaker to separate module
